+++ primefaces/src/main/java/org/primefaces/component/datatable/DataTableBase.java
@@ -67,6 +67,7 @@ public abstract class DataTableBase extends UIPageableData implements Widget, RT
         filterBy,
         filterDelay,
         filterEvent,
+        filterNormalize,
         filteredValue,
         frozenColumns,
         frozenColumnsAlignment,
@@ -785,5 +786,14 @@ public abstract class DataTableBase extends UIPageableData implements Widget, RT
 
     public void setCellNavigation(boolean cellNavigation) {
         getStateHelper().put(PropertyKeys.cellNavigation, cellNavigation);
+    }
+
+    @Override
+    public boolean isFilterNormalize() {
+        return (Boolean) getStateHelper().eval(PropertyKeys.filterNormalize, false);
+    }
+
+    public void setFilterNormalize(boolean filterNormalize) {
+        getStateHelper().put(PropertyKeys.filterNormalize, filterNormalize);
     }
 }