+++ primefaces/src/main/java/org/primefaces/component/treetable/feature/FilterFeature.java
@@ -30,6 +30,7 @@ import java.util.List;
 import java.util.Locale;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicBoolean;
+
 import javax.el.ELContext;
 import javax.faces.FacesException;
 import javax.faces.context.FacesContext;