+++ docs/15_0_0/gettingstarted/whatsnew.md
@@ -32,6 +32,7 @@ Look into [migration guide](https://primefaces.github.io/primefaces/15_0_0/#/../
     * Added attribute `frozenColumnsAlignment` to support alignment of frozen columns left and right
     * Added `expandIcon`/`collapseIcon` attributes for RowToggler
     * Added `rowsPerPage` in the PageEvent AJAX event
+    * Added `filterNormalize` attribute to normalize the filter values (remove accents)
 
 * FeedReader
     * Added `podcast="true"` property if [Apple Itunes Podcast](https://help.apple.com/itc/podcasts_connect/#/itcb54353390) parsing and specific tags 
@@ -55,5 +56,8 @@ Look into [migration guide](https://primefaces.github.io/primefaces/15_0_0/#/../
 
 * Tree
     * Added `filterPlaceholder` property
+
+* TreeTable
+    * Added `filterNormalize` attribute to normalize the filter values (remove accents)